LayoutTests/webaudio/offlineaudiocontext-suspend-resume-graph-manipulation-expected.txt


PASS # AUDIT TASK RUNNER STARTED. 
PASS Executing "test" 
PASS Audit report 
PASS > [test] Synchronous graph manipulation with suspend() and resume() 
PASS   Frame at which context is suspended is equal to 12800. 
PASS   Connecting a constant buffer to destination and starting at 12800 frame did not throw an exception. 
PASS   Context is suspended is equal to 25600. 
PASS   Disconnecting a constant buffer at 25600 frame did not throw an exception. 
PASS   Buffer frame [0, 12800) contains only the constant 0. 
FAIL X Buffer frame [12800, 25600): Expected 1 for all values but found 12800 unexpected values: 
	Index	Actual
	[0]	0
	[1]	0
	[2]	0
	[3]	0
	...and 12796 more errors. assert_true: expected true got false
PASS   Buffer frame [25600, 38400) contains only the constant 0. 
FAIL < [test] 1 out of 7 assertions were failed. assert_true: expected true got false
FAIL # AUDIT TASK RUNNER FINISHED: 1 out of 1 tasks were failed. assert_true: expected true got false
